Revert "[SLES] Do not divid by zero."
This reverts commit 6293ac813834493c23d24446abc43a26639e7e75.

The solution used would only work with the Jacobi solver for scalars, so
if we have 0 entries on the diagonal due to porosity, this should be
handled upstream in a cleaner manner (setting 1 on the matrix diagonal,
and 0 on other terms).

The Jacobi solver includes a division by a diagonal. If the diagonal
is zero, true Jacobi shouls not be expected to handle it.

At the very least, this could be handled as an option, but not a standard
solver, risking a degradation of its performance (for example possibility
of vectorization here).