[DRSM inlet] improve inlet profiles for Rij-epsilon models.
A simply shear flow is imposed with an anisotropic value of
Rnt = -sqrt(Cmu) k
Direction n is given by the velocity direction, whereas the direction t
is given by the direction pointing to the (nearest) wall. Though it
require the computation of the distance to the wall (and its gradient).
The gradient could be computed only once (and stored).