GUI backwards compatibility: cleaner/more robust handling of version history.
The version history is no sorted to remove duplicates, though the version
currently used and back compatibility are always based on the last element
in case a cas is opened and modified using an older version than the one
used for creation (not recommended in any case, but we try to limit
damages here).

"-patch" extensions are also removed.