[Build] Update and improve prerequisite detection for recent SALOME versions.
Cherry picked from commits 4c6fb262 and 55d21332 on master.

This allows using 9.9, 9.10 and CAS builds for HDF5, MED, CGNS,
ParaView Catalyst.

This could allow the same builds for PT-SCOTCH, but the build options
of PT-SCOTCH 6  are currently incompatible with the default
code_saturne MPI_Init_thread configuration. Moving to PT-SCOTCH 7.0
may solve the issue in the future.