Improve/fix some C/Fortran mappings.
Many turbulence constants were not useful anymore in Fortran, but a few were
modified in the legacy Fortran version of modini based on turbulence model
options, where they should have been modified in C.

This is now fixed, and many variables not needed anymore in Fortran are
removed from the cstphy module.

Also, where possible, initialization of Fortran values with the 'save'
attibute is done directly in the declaration, rather than in iniini.f90.

A few other minor Fortran->C mappings are improved.