Further fixes relative to commit 72e901c5.
This reverts and adapts most of the tests on `unstd_multiplicator`,
which must be specific to each field and not global.
The "is_tepmerature" field keyword and "iscacp" may not be well
chosen anymore with the addition of internal-energy based schemes,
but should be renamed as a separate step, or better, replaced
by multiplier info in a field's cs_equation_params_t/var_cal_opt
structure (which would avoid testing separately for Cp and Cv).