[SLES] Deploy the usage of different tolerances: ATOL, DTOL, RTOL
This is the same way to set the stopping criteria as in the PETSc
library. Up to now, this is used with PETSc solvers and in some
specific saddle-point solvers.

atol = absolute tolerance
dtol = divergence tolerance
rtol = relative tolerance

CS_EQKEY_ITSOL_EPS is now deprecated. Please prefer the usage of
CS_EQKEY_ITSOL_RTOL