Internal coupling: first refactoring stage for least-squares or hybrid gradients.
For these gradient cases, we build local (temporary) coefa/coefb arrays using
the coupled field's hint/hext values. This allows handling the coupling
as a preparation/update of the coefa/coefb arrays, avoiding specific loops in
all algorithms.

In theory, we should use reconstructed values on the boundary for coupled values
excange, but when the gradient is stronger in the normal direction, to the
wall, on tetrahedral meshes (those most needing reconstruction), the low gradient
quality can degrade results instead of improving them. This will probably require
dedicated heuristics or options.