Improve on SYRTHES wrapper (commit 2c0eeff9).
The needed environment is sourced in the coupling mpmd script (and "script"
MPMD mode is forced) when a coupled domain needs to source a shell environment.
This allow using local tools (such as the debug wrapper in the standard
manned fo coupled cases, though this will need to be improved for parallel
tools such as ddt or vtune.