Improve control of mesh preprocessing in case of restart.
The behavior is now handled explicitely in the solver (and not only
the script) so as to improve compatibility with Cartesian mesh
definitions. The previous behavior was based only on the presence
of mesh_input files or directories.

The setting may also be forced using `cs_preprocessor_data_set_restart_mode`
in `cs_user_mesh_restart_mode`.