Multigrid pairwise aggregation: tentative fix for corner case.
In some cases, the update of the search set for an aggregated pair
could lead to the remaining adjacency size of the aggregated element being
reduced to 0 (a_m[j] = 0) for the first element of the pair.
In this case, we should simply continue instead of trying
to reduce it again when handling that aggregated element.

This should fix GitLab issue #506.