Allow setting of linear solver parameters for some CDO variables using GUI.
This is only a temporary solution, as it does not allow defining the wider
range of predefined solvers available for CDO, but at least ensures that if
non-default solver settings were defined with the GUI or the
`cs_user_linear_solvers` function, it is not rejected by comparison with
higher-level equation parameters settings.

In the future, the GUI should define values for equation parameters, and
those should in turb be used for detailed/low level linear solver setups.