Fix Github issue #103.
Cherry picked from commit c1ad857c on master.

The hybrid convection scheme (used for RANS / LES turbulence models) and the
TVD/NVD convection schemes could not be used with no slope test or with the
beta limiter inside the thermal convection diffusion balance.

A spurious check was preventing it whereas the schemes were correctly
implemented below in that case.