Fix Github issue #103.
The hybrid convection scheme (used for RANS / LES turbulence models) and the
TVD/NVD convection schemes could not be used with no slope test or with the
beta limiter inside the thermal convection diffusion balance.

An spurious check was preventing it whereas the schemes were correctly
implemented below in that case.