Use more consistent cast for PETSc KSP structures in user examples.
This cast is used to avoid moder compilers complaining about
incorrect argument types.

And we do not want to include PETSc headers except where absolutely
necessary (i.e. not in code_saturne headers), because this leads to many
compilation warnings which drown out our own warnings, to which we should
pay attention.