GUI: add activation of cooling towers model.
We may need to change the design of the Analysis Features view page,
and replace radiobuttons with checkboxes, to allow combining several
physical models, but this should be done only of we expect multiple
models to be compatible with each other, which is not usually the case.

An alternative solution here would be to use a secondary combobox
below the first one (as is done for combustion) to allow selecting the
compatible atmospheric flow type, and activate the atmospheric model
in the appropriate cases.