Add least-squares vector gradient algorithm with fixed-point boundary logic.
This is similar to the algorithms in cs_gradient_boundary, and uses
a fixed-point logic for the boundary values instead of a higher rank local
system inversion.

For tetrahedral meshes, this seems to be at least twice as fast
(for the boundary gradients). For orthogoanl meshes, a factor of at least
5 has been observed.

This algorithm is also simpler.

It is not activated by default (yet), but can be activated using the
following environment variable:
`CS_GRADIENT_LSQ_BOUNDARY=fixed_point`