Fix extra occurences of variables in outpout volumic variables list.
Nodes defining variables or properties should not be named 'scalar',
though this seemed to be the case for some variables in the past.

Otherwise, when using internal coupling, the temperature can appear
both as a vraible and under a scalar node, leading to a double
appearance in the view.

In all cases, 'scalar' is used in many other places (such as defining BC's),
so should not be used in the same sense as 'variable', 'property',
or 'time_average'. Otherwise, we would need complex rules to handle
the XML file rather than simply loop over all nodes of a given sort.