Major commit: use unit normal instead of face normal for many operators.
The long-term goal is to remove the non-unit normal from the mesh quantities
structure, so as to save memory. Using the unit normal requires a few
additional multiplications but also saves divisions, and in the case of
porous media, only the surface needs to have a different value.
Also, the unit normal could probably be handled in single precision
in the future, and is a good candidate for this in any case.