Remove cs_mesh_b_faces_thread_block_range function.
This function assumed boundary faces are sorted by associated cell
adjacency, which is usually the case, but could depend on other boundary
face numberings. In addition, it is not adapted to dynamic scheduling.
Replacing the single (upt to now) use of this function by a simple
loop on boundary cells, the obtained performance on a CPU seems almost
identical (often  a few percent slower but sometimes slightly faster, given
measurement noise).