Fix CFL computation with zero fluid-volume cells.
In commit 5ac36a5d (2023-12-12), we attempted to use the fluid volume
instead of the mesh volume for more relevant CFL computations in
immersed boundary and VoF computations, but some divisions were not
handled correctly.

Another solution (which would avoid an arbitrary minimum cell size,
would be to check for NaN values locally, and replace those
when both numerator and denominator of a sub-expression are zero.
Here we simply assume cells cannot have a fluid volume lower than
1e-24. At this size, we should be far from the Navier-Stokes hypothesis.