CUDA linear solvers: only create new stream if default stream assigned to matrix.
This will allow avoiding to create additional streams in the context of
a multigrid solver. An option to allow creating specific streams or using
stream 0 only at the cs_sles_it_t or cs_sles_t level may be of use in the
future, but for now this mechanism should work.