CUDA: Add CUDA-version FCG as AMG smoother.
Based on patch by Li MA, with some simplifications (no need for a dedicated
smoother version of the FCG, as we can avoid the convergence test by
using a negative precision in the caller, and change from commit ba96d888
allows avoiding creating a new stream in the context of a smoother
called by a multigrid cycle.

TODO: in the non-cuda version, we should also use the standard FCG solver
instead of the local PCG here.