Wall distance computation: automate handling in case of restart.
We now automatically detect whether the wall faces have changed in
case of restart, so avoid recomputation when the wall distance and
its associated boundary conditions have been read and the mesh is fixed.
This removes the need for specifica handling for turbulence models such
as k-omega, with the positive/negative icdpar values even most old-timers
have forgotten or never learned about.

Using the porous model, this automation can not be handled correctly,
as we do not have all the available info, so recomputation is forced
for safety.