Allow CUDA nvcc to be used as C++ compiler, with gcc backend.
This allows using the dispatch class "as intended".
Prior to this, using dispatch in both .cu and .cpp files, with a different
compiler for each (nvcc and g++), the device context would not work properly
and seem to contain inconsistent values when viewed under a debugger,
while it seems to work fine if used with a single compiler. We suspect
this is due to the way the compilers avoid duplicating templated code.