Continue cleaning buoyant turbulent production terms. It fixes commit 6ed56f70 (2024-02-23)
Now, thermal expansion is computed when computing EOS so that, for
atmospheric flows, we always define beta.
It is now correctly computed for humide atmosphere (at least coherently
for k-eps and Rij-eps).
For Rij-eps, we still miss the buoyant terms due to humidity (only
temperature matters, and beta=1/theta_v)
MAJOR CHANGE Buoyant terms for atmosphere are now using current thermal
field (which is coherent with the EOS). This mainly impact all humide
cases.