Add base SYCL support.
Tested on CPU, not yet on GPU.

Mapping the dispatcher to SYCL's parallel_for required removing the
second lambda function argument (sum type) for scatter loops on interior
faces, so an additional `get_parallel_for_i_faces_sum_type` member
function of `cs_dispatch_context` was added to compensate for this.
Note that we would have had the same issue/requirement for mapping
to Kokkos.

The memory system also reports some non-freed memory blocks, but this
is because some host-only memory allocations are freed using `bft_mem_free`.
Using the same map for host-device and base memory allocations should
fix this (it will need to be tested to ensure the overheas associated
with maintaining the map for all allocations is nto too high, but the
overhead we had in older versions when always maintaining an allocated
arrays list was minimal. We did not use a map, but searched from the tail
of the array, so that could also be a viable option.