Add missing dispatch wait calls.
Also, use temporary arry for upwind face counts in
cs_convection_diffusion_vector, so as to prepare for general dispatch.
This will be a bit more costly when logging is required, but les costly in
the more common case where verbosity is not increased, as we do not then
need a reduction.