Fix extra preprocessing for staged restarts.
Cherry picked from commits 5643c02e, 90451b87, and 35ce78f9.

When a restart was to use an unmodified mesh ("restart/mesh_input.csm"),
the case domain's self.mesh_input entry was set in the "stage" step,
but not restored in the "preprocess" step in case of a 2-step submit,
leading to extra preprocessing which could lead to inconsistent
restart errors.

We now check this case so as to avoid unwanted preprocessing.

When restarting with no mesh change, add previous mesh_input to checkpoint.

This fixes GitLab issue #579.