GPU: minor optimizations based on testing on A100.
The memadvise calls have not been check yet for performance, so
are "placeholders".

On a 12 million cell case using 16 CPU cores and an A100 GPU, a large time
was observed for initializing the vector gradient reconstruction to zero.
Given that the "gather version" does not need to pre-initialize each term
to zero, but can do this directly in the main (interior faces, or cell/cells)
kernel, as it works cellwise, avoiding this extra initialization now makes
the algorithm slightly faster using the cellwise gather approach.