Fix extra preprocessing for staged restarts. starts.
When a restart was to use an unmodified mesh ("restart/mesh_input.csm"),
the case domain's self.mesh_input entry was set in the "stage" step,
but not restored in the "preprocess" step in case of a 2-step submit,
leading to extra preprocessing which could lead to inconsistent
restart errors.
    
We now check this case so as to avoid unwanted preprocessing.
    
This fixes GitLab issue #579.