Multigrid and linear solver optimization when initial solution is zero.
In many cases, this avoids an initial SpMV product, and can also avoid
requiring to initialize the solution to zero, which has no computational
intensity. On an example run on a DGX-A100, with about 3 million cells
per GPU, this lowers the multigrid V-cycle time from 163 to 146 seconds
(results may be a bit noisy), for an improvement between 10 and 15%.

This is currently only activated inside the multigrid solvers.
Using this for general solvers will require a second small change.
It should improve performance especially in cases with small time
steps where only a few iterations are needed for solvers, so that
a gain in the fits iteration can have noticeable impact.