Add `parallel_for_reduce_sum` method to `cs_dispatch`.
As there is no single naming scheme for this type of operation
(transform_reduce in standard C++, parallel_for with reduction argument
in SYCL, parallel_reduce in Kokkos), we choose a simple name
for the most commn operation types.

Although not as elegant as adding a "reduce" argument to a
parallel_for function, this will be easy to "grep" if we later decide
to rename or refactor it.