GPU: use gather algorithm for slope test gradients on GPU.
Here, the gather version of the algorithm is about 25% faster for vector
gradients than the scatter version (on an NVIDIA GeForce RTX 4060 Laptop GPU,
for the BUNDLE_C016_04 test case with 4 OpenMP threads), but the scatter
version is faster on CPU, so we use a different dispatch option on each.

Also, to avoid compiling extraneous versions, we use static casts
of the cs_dispatch_context class to its componant cs_host_context
and cs_devince_context.