Better handling of "restrict" qualifier, to alloc C++ __restrict__ extension.
The previous restrict qualification for pointers was ignored/lost when moving
some files to C++, and the "a[restrict]" syntax cannot be used with the
__restrict__ extension of gcc, clang, or nvcc. "*__restrict__ a" must be
used instead. For multidimensional arrays, "a[restrict][3]" for example
must be replaced by "(*restrict a)[3]".

Also, the restrict qualifier is to be used for local pointers, rather than
for the pointer-to object, so using it in function headers is not useful
(except for inline functions). So some additional cleanup is done here.