Multigrid: use optimized smoothers at all levels, but only if consistant.
If a optimized smoother is not available for a requested solver type,
or a convergence test is requires, default to regular solvers.

Since commit 4cc7d077 (2019-04-05), smoothers requiring a convergence test
in addition to a maximum iteration count (not recommended) were not handled
correctly, as they would reach the maximum itetation count.