CUDA: fix compilation with AMGX.
Following 1929fb50, a special case for the CUDA compiler is removed.
This breaks compilation with CUDA 9.0, so regarding compatibility with older
tools, we must choose between CPU and GPU toolchains. As there are
other issues/limitations with the current CPP dispatch code and older
nvcc versions, we drop support for CUDA 9.0.