Move first test on control_file after checkpoint read.
When loading a state with an ALICES FMU, the current state is also
immediately queried, meaning that field values must be allocated
and initialized.

This should otherwise change the previous behavior ony if some initalization
functions depend on the maximum time step value, which should never be
the case (being bad practice).