Restore update of L2 residuals.
In the future, this should be done with a time control setting independent
of that of the main log (though similar options should be provided).
Also, the computation should be improved, so as to use vector norms
for vectors (rather than independent components), and for better
efficiency, but since this is not updated at each time step by default,
it is not urgent.