Multigrid coarsening: threaded version for values assembly with MSR structures.
The top-level "ax0" term stil needs to be adapted, as well as the
convection-diffusion variant.

During this intermediate step, setting the `CS_MG_GATHER` environment
variable is needed to active the gather-based algorith. This variable is
not otherwise documented as its role is purely temporary, and it will be
removed when the algorithm path is complete.