Multigrid: avoid storing xa0ij for finest grid in most cases.
In the convection-diffusion case, this is still computed in the same
place as before.

Also, for MSR matrices with symmetric coefficients, the xa0 term does
not need to be mapped to a "native" (face-based) coefficients array,
so it should be possible to use a matrix directly assembled to MSR.

For non-symmetric matrices, the legacy face-based array is still needed,
because averaging Aij and Aji is difficult when one of i or j is in
an extended column (i.e. ghost region), as extended rows are not built into
an MSR matrix.