Lagragian 2-way coupling: remove unused coal combustion parameters.
Note that the cpgd1, cpgd2, and cpght coal devolatilization values are
still computed by the coal model, but they were only added to Lagrangian
2-way coupling in an older coal+Lagrangian model (CS_COMBUSTION_PCLC / icpl3c),
which was removed in c541b149 (2023-09-06). We assumed they are not needed
in the newer modelds, but they might simply not have beed called due to a bug.
More test cases would be needed for "coal + Lagrangian" cases.