Assemble matrix coefficients directly when possible (work in progress).
This is now done for the pressure matrix, but should be extended
to other matrix types, to remove an unnecessary copy and conversion,
and use a gather algorithm (with simpler parallelism) to compute
matrix coefficients. This should improve GPU and multithreaded
performance, while not significantly degrading sing thread CPU
performance (as some of the extra cost is some steps is offset by
removing another step).