Ensure cs_parall_thread_range does not return an out of bounds start id.
When the number of threads is high compared to the data size, even if
this is not recommended, alignement of the start index could move it past
the end index. Although this is not an issue if we loop directly between
the start and end index (as start is higher than end, the loop is empty),
When we use an indirection, such as index[start_id], index[end_id],
we could have an out of bounds access.