[Lagr] Estimate relaxation time for the fluid temperature based on lagr_time
An additional turbulent constant cs_turb_crij_ct is stored for this
relaxation time so it can be modified by user.
Note: if the variance of carrier temperature is computed, estimate the value of
this constant in coherence with the choice made for the variance_dissipation
constant.
Note also we simplify the temperature integration.