[Atmo] continue default setting of reference value (t0, ro0, p0) with meteo files.
The value of the GUI for ro0 is also not read if predefined law on.
This point has to be checked. It would be better to remove default
writting of `initial_value` for density when not necessary (e.g. when
metoe file is used).

Warning: this will change default values of some validation cases
because p0 and t0 was not considered when using meteo files.