GUI and XML model: cleanup references to some inactive models in XML.
Ideally, all references to specific models should be removed from
the XML structure when they are not active, but this requires
cleaning up to avoid crashes in unchecked accesses to the corresponding
nodes. This can be impreoved over time, but an overhaul of the model
structure so as to allow "ghost" nodes allowing easy access but not
output in the XML would be better. Using a hidden '_model' attribute
could work, assuming we can also cleanup the full subtree when the model
is 'off'; this assumes we document things a bit more...

Also, cleanup referencing some models is done when activating another
model, which involves cross references. It would be much more consistent
to have a pointer to the of current model's "set" function,
and call the corresponding call to turn it off when activating
another model.