GPU: add device memory pool to avoid costly memory free.
Part of the logic (number of tries for a match before a block is
evicted from the pool and freed)  is similar to code written
by D. Dubuc (ANEO), but in the this variant, the pool is used only to
transfer memory that should normally be freed, and can be activated,
deactivated, or parametrized on the fly.

The overall chages to the code are relatively minor, and some of the
previous code was refactored.

Some minor statistics are also available (peak size) and this could be
improved in the future.