Fix test on fixed-point requirement for boundary_gradient.
With fields with zero initial value and BC coeff a 0, the reconstruction
was not done even when needed. The new test is simply based on
BC coeff's B values.

This "standanlone" gradient is currently used for boundary
conditions, and some postprocessing and internal coupling, but
should be the basis for precomuting boundary values in the future.