[GUI] added turbulence initialization in neptune_cfd gui.
Similar to what is done in code_saturne gui, choice between
initialization from reference values and a formula for each field.

The model RpRfp is not done yet. It would require creating the tensor
Rfp instead of having to define the qfpxx, qfpyy, etc...

The possibility to choose the k-omega model is also added in this
feature.