Remove arrays related to legacy izfppp.
Note that for gas combustion, BC data previously stored in checkpoint files
should not be needed when restarting, as it seems it can be computed
directly from the BC definitions.

EBU and LW model initialization is temporarily broken by this commit,
as the mean inlet temperature and mixture fraction values need to be computed
with the C++ inlet zone info to replace a commmented-out section,
where qimp, fment, and tkent were prematurely removed.