Combustion (EBU and LW-models): complete missing initialization from 20b64e44.
Also remove specific turbulence initialization, so as to allow handling
of turbulence initialization in a standard way. THe previous initialization
used a very small reference velocity instead of "uref". For transient
cases, this may make sense, but allowing a "uref_ini" or changing uref
over time using the general turbulence model is more maintainable
and consistent.