Gas combustion: partial fix for division by 0 when not using JANAF.
Introduced in commit 807237dc (2014-11-20), where the fix can apply only
to the JANAF case, as the stoichiometric coefficients are not known otherwise.

The code later encounters a division by 0 in cs_combustion_gas_yg2xye,
as both cm->n_gas_el_comp = 0, and cm->coefeg = 0.

The "non-JANAF" mode seems to have been broken for a long time without
anybody noticing, so removing it might be a good idea.