Dispatch: add parallel_for_reduce for user-defined reduction on CPU.
The method still needs to be implemented for CUDA at least, but
a base API and unit test are now set up, and can be improved in
the future.

The wait() method is also now handled in the mixin the same as parallel_for
loops, so that it can also be used when directly using a host only context
(for consistency).