Try to simplify wall function a bit.
Done with Samuel Gaucher.
Function as _update_physical_quantities_smooth_wall should be removed and
merged with cs_wall_functions_velocity and all wall function
which take buoyancy into account (such as atmospheric wall functions)
should be rewritten in it.

Also "uiptn" can be removed, it makes the code unclear and is only used
in the log file.