EBU an LW gas combustion models: remove second initialization step.
This step actually amounted to a no-op (except halo syncs which could be done
earlier) for the LW model prior to conversionto C++ in commit a45148d1
(yesterday).
It was also the only instance of a second, non-user initialization in
the first solve pass (possibly overwriting user-defined values).
Since this was actually only used for the EBU model, untested to our
knowledge for more than 10 years (probably 15), it is better to remove
this, and let the user assign a specific initial condition (that will
not be overrriden by the code) if needed to accelerate convergence.

Also add a bit of cleanup, using field pointers in the combustion model
rather than searching for fields by name.