cs_solver should not have been renamed to .cpp in commit a88e9253.
This top-level file should be linked with the general C++ compiler, so
following our convention, should keep the .cxx extension.

The spurious change did not break the build because this file is linked
separately, with a Python script instead of directly with automake,
but did not follow the convention used elsewhere, so did break the code
with respect to the documentation.