Fix crash in coal model due to side-effect of Burke-Shchumann combustion model.
An additional coeff_therm array was defined as a global variable and
always filled, but was too small in the case of the coal model.
This array is now placed inside the combustion model structure rather
than as a global variable with no namespacing precaution, and used
only with the matching model.