Multigrid: other fixes for grid merging.
This reverts 0981c402, which caused a different issue.

The fix was redundant with an upstream filter, and led to a matrix
with a different norm in some cases, and degraded convergence.
so the original issue still needs to be fixed.

(cherry picked from commit 7d91c37c6a81af8d4ff162557f7f23115ee3993c)