Halo exchange: excess read lenght in strided case, from b5171baf (2021-07-20).
the stride was applied twice. It seems that since the send lenght was correct,
although the message length requested when reading was too long, this generated
a "silent" error (visible in MPI status details, but not fatal) with no
additional consequences in most cases, but fixing this is of course safer.