Multigrid (bottom grid merge): yet another fix following a9bea6ed (2025-05-19).
Also enable bottom grid merging in single-rank mode in presence of
periodicity, so as to allow merging of halo columns into the main matrix.
This may save a bit of latency due to function calls or especially GPU
kernel launches, but the main advantage is that the bottom-level matrix
is made square, and the system could then be solved with non-distributed
linear solvers (including direct solvers from vendor libraries).