Time moments: use reference time step instead of local time step.
This make no difference at all for constant or adaptive time steps,
and simplifies the code logic when using a local (pseudo-steady) time step.
In that case, the average cannont be interpreted as a true time average
anyways, only as an "ensemble" mean (with a smoother time behavior than
the matching variables), so using a simpler logic seems as good as the
previous one.

Compatibility is ensured when restarting from a older checkpoint
file (still using a weight accumulator based on the matching mesh
location, initialized with older local values and updated with
the uniform reference time step.