CUDA: implement fused block Jacobi kernels (not used by default).
This requires syncing threads, using an additional shared memory array,
and may modify alignment of array accesses as only a number of threads
equal to a multiple of the matrix diagonal block size per CUDA block
may be used, shifting addresses accessed by following threads.

On a workstation, this does not seem to change performance, so
the fused kernels are not activated at this stage, pending further
tests on clusters.

CUDA: avoid some unneeded operations in Jacobi solver when residual not needed.

t