CUDA: fixes and maintainability improvements (use of parallel dispatch) for GCR.
The GCR algorithm had not changed recentyly, but issues with convergence
only revealed themselves recently. This might be an effect of a latent
bug in reduction block/grid sizes or synchronizations.

Switch to dispatch / lambda-functions allows better alignment with
thr CPU version, better readability, and fixing of the observed
convergence issues.