[Atmo] Add a TODO for humide atmosphere init.
Thermal laws and initialization should be coherent to circumvent any
issues.
Moreover, the hydrostatic_pressure field and the meteo_pressure field
should be merged and computed the same way. Mind that
hydrostatic_pressure field is not really the hydrostatic part because it
does not contain rho0.g.z if I remember well.