Postprocessing: switch to cs_function_t-based output for MPI rank.
This allows easier user control over activation or deactivation
of this visualization.

As this changes the order under which field data is output, when opening
a case under ParaView (with the EnSight gold format at least), the
first (default) field is now the same whether we are running in parallel
or not, as the mpi_rank_id comes later.