Simplify headers inclusion order in user-defined functions.
If cs_headers is included first, we do not need to separately include
base/cs_defs.h. Note that this also requires including that file
first in the sequence to ensure cs_config.h is included before
any #if defined(HAVE_) macro. This was done indirectly in the calling
sequence, but is made more explicit here.