[Rad] add an option to save radiance by direction.
save_radiance_dir is set to true by default for atmo as it allows to
take diffusion from one direction to another into account.
Also, it saves time because the solving does not start with 0.

Some cleaning is also done with dedicated BC functions for DOM and P1.

For atmo, the delta redistribution is coded for the moment for solar
bands.