Wall functions: major commit, merge rough and smooth wall functions.
It namely clean atmospheric wall functions (Monin--Obukhov, Louis).
As buoyant parameter is now computed only once, it should be coherent
for all atmospheric wall functions even for humid atmosphere model.

Rough wall function can be removed now, as it is a sub-option of
wall models (icodcl = 5).
Therefore cs_wall_functions_2scales_log is also removed because
cs_wall_functions_2scales_smooth_rough give the same response for 0 roughness.

The run_solver.log file is also simplified as "uiptn" is now removed
(this value was not interpretable).

The rough-smooth wall function go back to the log ones when roughness
is zero (and not on the scalable one as before).