[Turb] rename local variable to improve readibility.
Some symmetric tensors which was taken as [j][i] for the component ij
are now taken as [i][j] (as it should be) even if this change is
transparent.
Production term, which is always computed, is now directly used for
epsilon (and no need to compute k_prod outside the loop).
Also for BFH Coriolis is now implicited as it should be.
Note that it should also be done for other model but this change is
let for another commit.